Move metric sets to metrics module
