+++ config-model/src/test/java/com/yahoo/vespa/model/admin/monitoring/MetricTest.java
@@ -1,6 +1,7 @@
 // Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 package com.yahoo.vespa.model.admin.monitoring;
 
+import ai.vespa.metrics.set.Metric;
 import com.google.common.collect.ImmutableMap;
 import org.junit.jupiter.api.Test;
