+++ metrics/src/main/java/ai/vespa/metrics/set/Metric.java
@@ -1,5 +1,5 @@
 // Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.vespa.model.admin.monitoring;
+package ai.vespa.metrics.set;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
