+++ metrics/src/main/java/ai/vespa/metrics/set/VespaMetricSet.java
@@ -1,5 +1,5 @@
 // Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.vespa.model.admin.monitoring;
+package ai.vespa.metrics.set;
 
 import ai.vespa.metrics.ClusterControllerMetrics;
 import ai.vespa.metrics.ContainerMetrics;
@@ -28,7 +28,7 @@ import static ai.vespa.metrics.Suffix.ninety_five_percentile;
 import static ai.vespa.metrics.Suffix.ninety_nine_percentile;
 import static ai.vespa.metrics.Suffix.rate;
 import static ai.vespa.metrics.Suffix.sum;
-import static com.yahoo.vespa.model.admin.monitoring.DefaultVespaMetrics.defaultVespaMetricSet;
+import static ai.vespa.metrics.set.DefaultVespaMetrics.defaultVespaMetricSet;
 import static java.util.Collections.singleton;
 
 /**
